replace options class with builder to simplify api
