+++ src/types.ts
@@ -1,3 +1,31 @@
+export interface ImgproxyConfig {
+  /**
+   * The url where imgproxy runs.
+   */
+  baseUrl: string;
+  /**
+   * If urls should be encoded with base64.
+   */
+  encode?: boolean;
+  /**
+   * If insecure usage is supported by the service, true by default.
+   * When set to a string it will be used as the plain "signature".
+   */
+  insecure?: boolean | string;
+}
+
+export interface ImgproxySecureConfig extends ImgproxyConfig {
+  /**
+   * The key to use for creating the signature.
+   */
+  key: string;
+  /**
+   * The salt to use for creating the signature.
+   */
+  salt: string;
+  insecure: false;
+}
+
 export type ResizingType = 'fit' | 'fill' | 'crop';
 
 export type HexColor = string;
