+++ tsconfig.json
@@ -3,7 +3,8 @@
     "lib": ["esnext"],
     "noUnusedLocals": true,
     "noImplicitAny": true,
-    "outDir": "dist"
+    "outDir": "dist",
+    "esModuleInterop": true
   },
   "include": [
     "src/**/*"
